Move ObjectClient to different package
